+++ multimodal/src/autogluon/multimodal/constants.py
@@ -93,7 +93,7 @@ METRIC_MODE_MAP = {
     MAP: MAX,
     MEAN_AVERAGE_PRECISION: MAX,
     OVERALL_ACCURACY: MAX,
-    RSUM: MAX,
+    HIT_RATE: MAX,
     NDCG: MAX,
 }
 VALID_METRICS = METRIC_MODE_MAP.keys()
