+++ multimodal/src/autogluon/multimodal/optimization/utils.py
@@ -14,18 +14,6 @@ from transformers import Adafactor
 from transformers.trainer_pt_utils import get_parameter_names
 import numpy as np
 
-try:
-    import torch.distributed.nn
-    from torch import distributed as dist
-    has_distributed = True
-except ImportError:
-    has_distributed = False
-
-try:
-    import horovod.torch as hvd
-except ImportError:
-    hvd = None
-
 from ..constants import (
     ACC,
     ACCURACY,
@@ -179,7 +167,7 @@ class CustomHitRate(torchmetrics.Metric):
         query_embeddings = torch.cat(self.query_embeddings)
         response_embeddings = torch.cat(self.response_embeddings)
         if self.logit_scale:
-            logit_scale = torch.mean(self.logit_scale)
+            logit_scale = torch.cat(self.logit_scale).mean()
         else:
             logit_scale = 1
 
@@ -269,7 +257,7 @@ def get_metric(
             None,
         )  # This only works for detection where custom_metric is not required for BaseAggregator
     elif metric_name == HIT_RATE:
-        return None, compute_rsum
+        return CustomHitRate, None
     else:
         raise ValueError(f"Unknown metric {metric_name}")
 
@@ -861,7 +849,7 @@ def get_matcher_loss_func(
             distance=get_metric_learning_distance_func(distance_type),
         )
     elif loss_type.lower() == MULTI_NEGATIVES_SOFTMAX_LOSS:
-        return MULTI_NEGATIVES_SOFTMAX_LOSS(
+        return MultiNegativesSoftmaxLoss(
             local_loss=True,
             gather_with_grad=True,
             cache_labels=True,
@@ -980,49 +968,3 @@ def compute_probability(
 
     return prob
 
-
-def gather_features(
-        image_features,
-        text_features,
-        local_loss=False,
-        gather_with_grad=False,
-        rank=0,
-        world_size=1,
-        use_horovod=False
-):
-    assert has_distributed, 'torch.distributed did not import correctly, please use a PyTorch version with support.'
-    if use_horovod:
-        assert hvd is not None, 'Please install horovod'
-        if gather_with_grad:
-            all_image_features = hvd.allgather(image_features)
-            all_text_features = hvd.allgather(text_features)
-        else:
-            with torch.no_grad():
-                all_image_features = hvd.allgather(image_features)
-                all_text_features = hvd.allgather(text_features)
-            if not local_loss:
-                # ensure grads for local rank when all_* features don't have a gradient
-                gathered_image_features = list(all_image_features.chunk(world_size, dim=0))
-                gathered_text_features = list(all_text_features.chunk(world_size, dim=0))
-                gathered_image_features[rank] = image_features
-                gathered_text_features[rank] = text_features
-                all_image_features = torch.cat(gathered_image_features, dim=0)
-                all_text_features = torch.cat(gathered_text_features, dim=0)
-    else:
-        # We gather tensors from all gpus
-        if gather_with_grad:
-            all_image_features = torch.cat(torch.distributed.nn.all_gather(image_features), dim=0)
-            all_text_features = torch.cat(torch.distributed.nn.all_gather(text_features), dim=0)
-        else:
-            gathered_image_features = [torch.zeros_like(image_features) for _ in range(world_size)]
-            gathered_text_features = [torch.zeros_like(text_features) for _ in range(world_size)]
-            dist.all_gather(gathered_image_features, image_features)
-            dist.all_gather(gathered_text_features, text_features)
-            if not local_loss:
-                # ensure grads for local rank when all_* features don't have a gradient
-                gathered_image_features[rank] = image_features
-                gathered_text_features[rank] = text_features
-            all_image_features = torch.cat(gathered_image_features, dim=0)
-            all_text_features = torch.cat(gathered_text_features, dim=0)
-
-    return all_image_features, all_text_features
